ha: u128 freshness compare + derive test threshold (#2332)
Fold the two Copilot nits on PR #2340 (both low/defensive; the freshness
logic and test-failover 14/0 are unchanged):

- heartbeat_fresh_mono compared a u64 age against
  HEARTBEAT_STALE_AFTER.as_nanos() (u128) via `as u64`. The 5s threshold
  is far below u64::MAX ns so no truncation can occur today, but compare
  in the u128 domain so the bound stays exact if the threshold ever grows.

- The freshness unit tests hard-coded 5_000_000_000 ns. Derive HB_STALE_NS
  from HEARTBEAT_STALE_AFTER.as_nanos() so the tests can never drift from
  the production threshold.

No behavior change; build + ha/heartbeat suite green.

+++ userspace-dp/src/afxdp/bpf_map/ha.rs
@@ -171,5 +171,8 @@ pub(in crate::afxdp) fn heartbeat_fresh_mono(last_heartbeat_ns: u64, now_mono_ns
         return false;
     }
     let age_ns = now_mono_ns.saturating_sub(last_heartbeat_ns);
-    age_ns <= HEARTBEAT_STALE_AFTER.as_nanos() as u64
+    // Compare in the u128 domain so the threshold can never be silently
+    // truncated by an `as u64` cast (HEARTBEAT_STALE_AFTER is 5s today, far
+    // below u64::MAX ns, but this keeps the bound exact if it ever grows).
+    u128::from(age_ns) <= HEARTBEAT_STALE_AFTER.as_nanos()
 }
